Reorganizing utilities.
1. Splitting `tb_util.js` into `tb_codegen.js` (for code generation) and `tb_runtime.js` (used when blocks are running).
2. Moving two callback helpers defined in `index.html` into `tb_gui.js`.
